# openapi.yaml
openapi: 3.0.3
info:
  title: GOD CARES 365 API
  version: 1.0.0
  description: DRF schema for Content, Discipleship, Notifications, Progress, Core,
    Shop, Mentorship
paths:
...
  /api/v1/content/categories/:
    get:
      operationId: v1_content_categories_list
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: query
        name: name
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: slug
        schema:
          type: string
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCategoryList'
          description: ''
    post:
      operationId: v1_content_categories_create
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
          description: ''
...
  /api/v1/content/posts/:
    get:
      operationId: v1_content_posts_list
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: query
        name: category
        schema:
          type: integer
      - in: query
        name: category_slug
        schema:
          type: string
      - in: query
        name: featured
        schema:
          type: boolean
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - in: query
        name: published_at
        schema:
          type: string
          format: date-time
      - in: query
        name: published_at__gt
        schema:
          type: string
          format: date-time
      - in: query
        name: published_at__gte
        schema:
          type: string
          format: date-time
      - in: query
        name: published_at__lt
        schema:
          type: string
          format: date-time
      - in: query
        name: published_at__lte
        schema:
          type: string
          format: date-time
      - in: query
        name: q
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
          enum:
          - archived
          - draft
          - published
        description: |-
          * `draft` - Draft
          * `published` - Published
          * `archived` - Archived
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPostListList'
          description: ''
    post:
      operationId: v1_content_posts_create
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetail'
          description: ''
  /api/v1/content/posts/{id}/:
    get:
      operationId: v1_content_posts_retrieve
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this post.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetail'
          description: ''
    put:
      operationId: v1_content_posts_update
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this post.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostDetailRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetail'
          description: ''
    patch:
      operationId: v1_content_posts_partial_update
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this post.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedPostDetailRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedPostDetailRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedPostDetailRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetail'
          description: ''
    delete:
      operationId: v1_content_posts_destroy
      description: Kwa GET bila staff, rudisha tu yaliyopublished endapo model ina
        is_published.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this post.
        required: true
      tags:
      - v1
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
...
    CategoryRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: ^[-a-zA-Z0-9_]+$
        description:
          type: string
      required:
      - name
...
    PaginatedCategoryList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Category'
...
    PaginatedPostListList:
      type: object
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/PostList'
...
    PatchedPostDetailRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
        content:
          type: string
          minLength: 1
        excerpt:
          type: string
          maxLength: 300
        featured_image:
          type: string
          format: binary
          nullable: true
          pattern: (?:jpg|jpeg|png|webp|gif)$
        status:
          $ref: '#/components/schemas/StatusDaeEnum'
        featured:
          type: boolean
        views:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        category:
          type: integer
        author:
          type: integer
...
    PostDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
        content:
          type: string
        excerpt:
          type: string
          maxLength: 300
        featured_image:
          type: string
          format: uri
          nullable: true
          pattern: (?:jpg|jpeg|png|webp|gif)$
        status:
          $ref: '#/components/schemas/StatusDaeEnum'
        featured:
          type: boolean
        views:
          type: integer
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        published_at:
          type: string
          format: date-time
          nullable: true
        category:
          type: integer
        author:
          type: integer
      required:
      - author
      - category
      - content
      - created_at
      - id
      - title
      - updated_at
    PostDetailRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
        content:
          type: string
          minLength: 1
        excerpt:
          type: string
          maxLength: 300
        featured_image:
          type: string
          format: binary
          nullable: true
          pattern: (?:jpg|jpeg|png|webp|gif)$
        status:
          $ref: '#/components/schemas/StatusDaeEnum'
        featured:
          type: boolean
        views:
          type: integer
        published_at:
          type: string
          format: date-time
          nullable: true
        category:
          type: integer
        author:
          type: integer
      required:
      - author
      - category
      - content
      - title
    PostList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 200
        slug:
          type: string
          maxLength: 200
          pattern: ^[-a-zA-Z0-9_]+$
        content:
          type: string
        excerpt:
          type: string
          maxLength: 300
        featured_image:
          type: string
          format: uri
          nullable: true
          pattern: (?:jpg|jpeg|png|webp|gif)$
        status:
          $ref: '#/components/schemas/StatusDaeEnum'
        featured:
          type: boolean
        views:
          type: integer
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        published_at:
          type: string
          format: date-time
          nullable: true
        category:
          type: integer
        author:
          type: integer
      required:
      - author
      - category
      - content
      - created_at
      - id
      - title
      - updated_at
...
    StatusDaeEnum:
      enum:
      - draft
      - published
      - archived
      type: string
      description: |-
        * `draft` - Draft
        * `published` - Published
        * `archived` - Archived
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT